requirements.txt
Django
celery
redis
python-dotenv
djangorestframework
dj-database-url
psycopg[binary]
langchain
langchain-google-genai
langchain-core
pydantic
requests